feature: move `Ensure` method from `ResultExtension` class to `Result` structure
