+++ test/unit/Global.cs
@@ -1,5 +1,4 @@
 global using Daht.Sagitta.Core.Monads;
-global using Daht.Sagitta.Core.Monads.Extensions;
 global using Daht.Sagitta.Core.UnitTest.Monads.Asserters;
 global using Daht.Sagitta.Core.UnitTest.Monads.Fixtures;
 global using Daht.Sagitta.Core.UnitTest.Monads.Mothers;
